move: dto 구조 변경
