+++ delivery-service/src/main/java/com/oneforlogis/delivery/application/dto/request/DeliveryAssignRequest.java
@@ -1,4 +1,4 @@
-package com.oneforlogis.delivery.application.dto;
+package com.oneforlogis.delivery.application.dto.request;
 
 public record DeliveryAssignRequest(
