+++ delivery-service/src/main/java/com/oneforlogis/delivery/application/dto/request/DeliveryRouteRequest.java
@@ -1,4 +1,4 @@
-package com.oneforlogis.delivery.application.dto;
+package com.oneforlogis.delivery.application.dto.request;
 
 import com.oneforlogis.delivery.domain.model.DeliveryRouteStatus;
 import jakarta.validation.constraints.NotNull;
